# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NGO Management System API",
    "description": "Comprehensive API for managing NGO operations, donations, donors, projects, and reports.",
    "version": "1.0.0",
    "contact": {
      "name": "NGO Management System Support",
      "email": "support@ngomanagementsystem.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://ngo-api.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/docs.json": {
      "get": {
        "tags": ["Core"],
        "summary": "Get OpenAPI Documentation",
        "description": "Returns OpenAPI specification in JSON format",
        "operationId": "getDocsJson",
        "responses": {
          "200": {
            "description": "OpenAPI specification"
          },
          "404": {
            "description": "OpenAPI document not found"
          }
        }
      }
    },
...
    "/api/donors": {
      "get": {
        "tags": ["Donors"],
        "summary": "List All Donors",
        "description": "Get paginated list of all donors (admin/staff only)",
        "operationId": "getAllDonors",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "donorType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["individual", "organization"]
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of donors retrieved"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      },
      "post": {
        "tags": ["Donors"],
        "summary": "Create Donor Profile",
        "description": "Create a new donor profile",
        "operationId": "createDonor",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "phone", "address"],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "MongoDB user ID"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+2348012345678"
                  },
                  "address": {
                    "type": "string",
                    "example": "12 Charity Street, Lagos"
                  },
                  "donorType": {
                    "type": "string",
                    "enum": ["individual", "organization"],
                    "default": "individual"
                  },
                  "organizationName": {
                    "type": "string",
                    "description": "Required if donorType is organization"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Donor profile created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/donations": {
      "get": {
        "tags": ["Donations"],
        "summary": "List All Donations",
        "description": "Get paginated list of all donations (admin/staff only)",
        "operationId": "getAllDonations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "confirmed", "failed"]
            }
          },
          {
            "name": "donationType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["cash", "transfer", "cheque"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of donations"
          }
        }
      },
      "post": {
        "tags": ["Donations"],
        "summary": "Create Donation",
        "description": "Record a new donation",
        "operationId": "createDonation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "projectId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number",
                    "example": 25000
                  },
                  "currency": {
                    "type": "string",
                    "default": "NGN"
                  },
                  "donationType": {
                    "type": "string",
                    "enum": ["cash", "transfer", "cheque"],
                    "default": "transfer"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Donation created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List All Projects",
        "description": "Get paginated list of all projects",
        "operationId": "getAllProjects",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["planned", "ongoing", "completed", "cancelled"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects"
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create Project",
        "description": "Create a new project (admin only)",
        "operationId": "createProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Water Project"
                  },
                  "description": {
                    "type": "string"
                  },
                  "budget": {
                    "type": "number",
                    "example": 100000
                  },
                  "amountReceived": {
                    "type": "number",
                    "default": 0
                  },
                  "amountSpent": {
                    "type": "number",
                    "default": 0
                  },
                  "status": {
                    "type": "string",
                    "enum": ["planned", "ongoing", "completed", "cancelled"],
                    "default": "planned"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/reports/projects": {
      "get": {
        "tags": ["Reports"],
        "summary": "Projects Report",
        "description": "Get detailed projects report",
        "operationId": "getProjectsReport",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects report retrieved"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      }
    },
...
  }
}